+++ django_filters/fields.py
@@ -1,6 +1,8 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+from collections import namedtuple
+
 from django import forms
 
 from .widgets import RangeWidget, LookupTypeWidget
@@ -21,7 +23,7 @@ class RangeField(forms.MultiValueField):
             return slice(*data_list)
         return None
 
-
+Lookup = namedtuple('Lookup', ('value', 'lookup_type'))
 class LookupTypeField(forms.MultiValueField):
     def __init__(self, field, lookup_choices, *args, **kwargs):
         fields = (
@@ -36,4 +38,6 @@ class LookupTypeField(forms.MultiValueField):
         super(LookupTypeField, self).__init__(fields, *args, **kwargs)
 
     def compress(self, data_list):
-        return data_list
+        if len(data_list)==2:
+            return Lookup(value=data_list[0], lookup_type=data_list[1] or 'exact')
+        return Lookup(value=None, lookup_type='exact')