+++ django_filters/filters.py
@@ -11,14 +11,14 @@ from django.utils import six
 from django.utils.timezone import now
 from django.utils.translation import ugettext_lazy as _
 
-from .fields import RangeField, LookupTypeField
+from .fields import RangeField, LookupTypeField, Lookup
 
 
 __all__ = [
     'Filter', 'CharFilter', 'BooleanFilter', 'ChoiceFilter',
     'MultipleChoiceFilter', 'DateFilter', 'DateTimeFilter', 'TimeFilter',
     'ModelChoiceFilter', 'ModelMultipleChoiceFilter', 'NumberFilter',
-    'RangeFilter', 'DateRangeFilter', 'AllValuesFilter',
+    'RangeFilter', 'DateRangeFilter', 'AllValuesFilter', 'MethodFilter'
 ]
 
 
@@ -30,7 +30,7 @@ class Filter(object):
     field_class = forms.Field
 
     def __init__(self, name=None, label=None, widget=None, action=None,
-        lookup_type='exact', required=False, distinct=False, **kwargs):
+        lookup_type='exact', required=False, distinct=False, exclude=False, **kwargs):
         self.name = name
         self.label = label
         if action:
@@ -40,6 +40,7 @@ class Filter(object):
         self.required = required
         self.extra = kwargs
         self.distinct = distinct
+        self.exclude = exclude
 
         self.creation_counter = Filter.creation_counter
         Filter.creation_counter += 1
@@ -47,6 +48,7 @@ class Filter(object):
     @property
     def field(self):
         if not hasattr(self, '_field'):
+            help_text = _('This is an exclusion filter') if self.exclude else ''
             if (self.lookup_type is None or
                     isinstance(self.lookup_type, (list, tuple))):
                 if self.lookup_type is None:
@@ -56,23 +58,23 @@ class Filter(object):
                         (x, x) for x in LOOKUP_TYPES if x in self.lookup_type]
                 self._field = LookupTypeField(self.field_class(
                     required=self.required, widget=self.widget, **self.extra),
-                    lookup, required=self.required, label=self.label)
+                    lookup, required=self.required, label=self.label, help_text=help_text)
             else:
                 self._field = self.field_class(required=self.required,
-                    label=self.label, widget=self.widget, **self.extra)
+                    label=self.label, widget=self.widget,
+                    help_text=help_text, **self.extra)
         return self._field
 
     def filter(self, qs, value):
-        if isinstance(value, (list, tuple)):
-            lookup = six.text_type(value[1])
-            if not lookup:
-                lookup = 'exact'  # fallback to exact if lookup is not provided
-            value = value[0]
+        if isinstance(value, Lookup):
+            lookup = six.text_type(value.lookup_type)
+            value = value.value
         else:
             lookup = self.lookup_type
         if value in ([], (), {}, None, ''):
             return qs
-        qs = qs.filter(**{'%s__%s' % (self.name, lookup): value})
+        method = qs.exclude if self.exclude else qs.filter
+        qs = method(**{'%s__%s' % (self.name, lookup): value})
         if self.distinct:
             qs = qs.distinct()
         return qs
@@ -189,3 +191,38 @@ class AllValuesFilter(ChoiceFilter):
         qs = qs.order_by(self.name).values_list(self.name, flat=True)
         self.extra['choices'] = [(o, o) for o in qs]
         return super(AllValuesFilter, self).field
+
+
+class MethodFilter(Filter):
+    """
+    This filter will allow you to run a method that exists on the filterset class
+    """
+    def __init__(self, *args, **kwargs):
+        # Get the action out of the kwargs
+        action = kwargs.get('action', None)
+
+        # If the action is a string store the action and set the action to our own filter method
+        # so it can be backwards compatible and work as expected, the parent will still treat it as
+        # a filter that has an action
+        self.parent_action = ''
+        text_types = (str, six.text_type)
+        if type(action) in text_types:
+            self.parent_action = str(action)
+            kwargs.update({
+                'action': self.filter
+            })
+
+        # Call the parent
+        super(MethodFilter, self).__init__(*args, **kwargs)
+
+    def filter(self, qs, value):
+        """
+        This filter method will act as a proxy for the actual method we want to call.
+        It will try to find the method on the parent filterset, if not it defaults
+        to just returning the queryset
+        """
+        parent = getattr(self, 'parent', None)
+        parent_filter_method = getattr(parent, self.parent_action, None)
+        if parent_filter_method is not None:
+            return parent_filter_method(qs, value)
+        return qs